+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 - All notable changes to `Messenger` will be documented in this file.
 
+## [v0.15 (02-13-2021)]
+
+### Removed
+- StoreEditMessage listener.
+
+### Changed
+- Edit message action will store the edit history immediately after update message.
+
 ## [v0.14 (02-09-2021)]
 
 ### Added